requirements.txt
certifi==2021.10.8
click==8.1.3
Flask==2.1.2
flatbuffers==2.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
onnx==1.11.0
onnxruntime==1.11.1
onnxruntime-gpu==1.11.1
Pillow==9.1.0
protobuf==3.20.1
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
